Animation, added new sprites, Shooting is a little off
I have animated the movement for both protagonists and added the other sprites to the project. Shooting for the Male character offsets his position, and I am unsure why. The Female can move the Male, but the Male cannot move the Female, I am also unsure of why. If able to, attempt to fix the offsetting of the Male when shooting, and see if you can determine why they have different behaviors.